[#183] Refactored FileAccessor Into it's own Class
THis now clears up alot of space in the maps activity and now allows other activities to use the class without any issues either
